# web/frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CKA Labs FastAPI backend",
    "version": "0.1.0",
    "description": "OpenAPI snapshot of the FastAPI surface consumed by the Vue frontend (FE-004). The live spec is preferred (POST http://<host>:3000/openapi.json); this committed snapshot keeps `gen:api` fully offline. Shapes mirror web/server.py Pydantic models and route return values. NOTE: this snapshot is hand-authored because the platform host accepted the TCP connection but stalled before sending the response body (see .agents/FE-004-implementer-2/EVIDENCE.md)."
  },
  "servers": [{ "url": "/" }],
...
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "operationId": "authLogin",
        "tags": ["auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated; sets the HttpOnly cka_auth_token cookie",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/api/auth/me": {
      "get": {
        "operationId": "authMe",
        "tags": ["auth"],
        "responses": {
          "200": {
            "description": "Authenticated user session",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthMeResponse" }
              }
            }
          },
          "401": {
            "description": "No user session",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthMeUnauthenticated" }
              }
            }
          }
        }
      }
    },
    "/api/admin/users": {
      "get": {
        "operationId": "listAdminUsers",
        "tags": ["admin"],
        "responses": {
          "200": {
            "description": "All users (admin only)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AdminUsersResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Error" }
        }
      },
      "post": {
        "operationId": "adminCreateUser",
        "tags": ["admin"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateUserRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created user (admin only)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CreateUserResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Error" },
          "401": { "$ref": "#/components/responses/Error" }
        }
      }
    }
  },
  "components": {
    "responses": {
      "Error": {
        "description": "Error response",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorDetail" }
          }
        }
      }
    },
    "schemas": {
      "ErrorDetail": {
        "type": "object",
        "required": ["detail"],
        "properties": {
          "detail": { "type": "string" }
        }
      },
...
      "LoginRequest": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": { "type": "string" },
          "password": { "type": "string" }
        }
      },
      "LoginResponse": {
        "type": "object",
        "required": ["status", "username", "role"],
        "properties": {
          "status": { "type": "string" },
          "username": { "type": "string" },
          "role": { "type": "string" }
        }
      },
...
      "AuthMeResponse": {
        "type": "object",
        "required": ["username", "role", "authenticated"],
        "properties": {
          "username": { "type": "string" },
          "role": { "type": "string" },
          "authenticated": { "type": "boolean" }
        }
      },
      "AuthMeUnauthenticated": {
        "type": "object",
        "required": ["authenticated"],
        "properties": {
          "authenticated": { "type": "boolean" }
        }
      },
      "CreateUserRequest": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": { "type": "string" },
          "password": { "type": "string" },
          "role": { "type": "string", "default": "user" }
        }
      },
      "User": {
        "type": "object",
        "required": ["username", "role", "created_at"],
        "properties": {
          "username": { "type": "string" },
          "role": {
            "type": "string",
            "enum": ["admin", "user"]
          },
          "created_at": { "type": "string" }
        }
      },
      "CreateUserResponse": {
        "type": "object",
        "required": ["status", "user"],
        "properties": {
          "status": { "type": "string" },
          "user": { "$ref": "#/components/schemas/User" }
        }
      },
      "AdminUsersResponse": {
        "type": "object",
        "required": ["users", "total"],
        "properties": {
          "users": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/User" }
          },
          "total": { "type": "integer" }
        }
      }
    }
  }
}